Specify the SPE payoff set(s) of the truncated game(s) after a cap in period T. While we could specify a complete repeated game that is played after the cap, it also suffices to specify just an SPE payoff set of the truncated game of the after-cap state.

Specify the SPE payoff set(s) of the truncated game(s) after a cap in period T. While we could specify a complete repeated game that is played after the cap, it also suffices to specify just an SPE payoff set of the truncated game of the after-cap state.

rel_after_cap_payoffs(
  g,
  x = NA,
  U,
  v1 = NA,
  v2 = NA,
  v1.rep = NA,
  v2.rep = NA,
  x.T = NA
)

Arguments

g
a relational contracting game created with rel_game

x
The state(s) for which this after-cap payoff set is applied. If NA (default) and also x.T is NA, it applies to all states.

U
The highest joint payoff in the truncated repeated game starting from period T.

v1
The lowest SPE payoff of player 1 in the truncated game. These are average discounted payoffs using delta as discount factor.

v2
Like v1, but for player 2.

v1.rep
Alternative to v1. Player 1 lowest SPE payoff in the repeated game with adjusted discount factor delta*(1-rho). Will be automatically converted into v1_trunc based on rho, delta, and bargaining weight. Are often easier to specify.

v2.rep
Like v1.rep, but for player 2.

x.T
Instead of specifiying the argument x, we can specify as x.T a name of the after-cap state. This can be refereed to as the argument x.T in rel_state and rel_states

Value

Returns the updated game