Finds an optimal simple subgame perfect equilibrium of g. From this the whole SPE payoff set can be deduced.

Finds an optimal simple subgame perfect equilibrium of g. From this the whole SPE payoff set can be deduced.

rel_spe(
  g,
  delta = g$param$delta,
  tol.feasible = 1e-10,
  no.exist.action = c("warn", "stop", "nothing"),
  verbose = FALSE,
  r1 = NULL,
  r2 = NULL,
  rho = g$param$rho,
  add.action.labels = TRUE,
  max.iter = 10000,
  first.best = FALSE
)

Arguments

g
the game object

delta
The discount factor. By default the discount factor specified in g.

tol.feasible
Due to numerical inaccuracies, sometimes incentive constraints which theoretically should exactly hold, seem to be violated. To avoid this problem, we will consider all action profiles feasible whose incentive constraint is not violated by more then tol.feasible. This means we compute epsilon equilibria in which tol.feasible is the epsilon.

no.exist.action
What shall be done if no pure SPE exists? Default is no.exist.action = "warning", alternatives are no.exist.action = "error" or no.exist.action = "nothing".

verbose
if TRUE give more detailed information over the solution process.

r1
(or r2) if not NULL we want to find a SPE in a truncated game. Then r1 and r2 need to specify for each state the exogenously fixed negotiation payoffs.

rho
Only relevant if r1 and r2 are not null. In that case the negotiation probability.